Fix DBE detection in daily/weekly to filter out duplicate entries.
